refactor: 데이터타입과 접근제어자 변경
MenuValidator, MenuGroup에서 public 메소드를 private로 변경하고, BadgeType에서 double 변수를 int로 변경했습니다.

+++ src/main/java/christmas/config/BadgeType.java
@@ -13,7 +13,7 @@ public enum BadgeType {
         this.price = price;
     }
 
-    public static BadgeType determineBadgeByPrice(double price) {
+    public static BadgeType determineBadgeByPrice(int price) {
         if (price >= SANTA.price) {
             return SANTA;
         }
